Update imshow_all
Passing unkown kwargs such as shape to imshow seems to raise an error
with more recent versions of matplotlib. A good opportunity, to
provide and docuemnt the keyword parameters in a more explicit way.

Note I'm not to sure on the strange calculation of width for the
figsize. The width gets smaller the more rows there are and multplied
with 0.75 but only for more than one rows?